requirements.txt
Flask==0.11.1
gunicorn==19.6.0
requests
line-bot-sdk
beautifulsoup4
google
numpy
pythainlp
#https://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
#transformers==3.5.0
#sentence-transformers